Convert electron from JS to TS
