+++ env.d.ts
@@ -1,1 +1,5 @@
 /// <reference types="vite/client" />
+
+// Taken from the docs: https://www.electronforge.io/config/plugins/vite#hot-module-replacement-hmr
+declare const MAIN_WINDOW_VITE_DEV_SERVER_URL: string
+declare const MAIN_WINDOW_VITE_NAME: string
