+++ tsconfig.app.json
@@ -1,6 +1,6 @@
 {
   "extends": "@vue/tsconfig/tsconfig.web.json",
-  "include": ["env.d.ts", "frontend/**/*", "frontend/**/*.vue"],
+  "include": ["env.d.ts", "frontend/**/*", "frontend/**/*.vue", "electron/*.ts"],
   "exclude": ["frontend/**/__tests__/*"],
   "compilerOptions": {
     "composite": true,
